+++ src/test/java/ladder/domain/UserTest.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain;
 
-import ladder.domain.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
